refactor: split SmartUITestRunner, ComprehensiveUITestRunner, and lib.ts (#52)
SmartUITestRunner (586 LOC) split into:
- src/runners/smart/SmartElementFinder.ts – DOM-based element discovery
- src/runners/smart/SmartInteractionExecutor.ts – tab interaction logic
- src/runners/smart/index.ts – barrel exports
- src/runners/SmartUITestRunner.ts – thin facade (263 LOC)

ComprehensiveUITestRunner (481 LOC) split into:
- src/runners/comprehensive/UIFlowTester.ts – tab test methods + testTab()
- src/runners/comprehensive/index.ts – barrel exports
- src/runners/ComprehensiveUITestRunner.ts – thin facade (176 LOC)

lib.ts (590 LOC) split into:
- src/lib/ConfigurationLoader.ts – createDefaultConfig + loadConfiguration
- src/lib/ScenarioLoader.ts – loadTestScenarios + filterScenariosForSuite + TEST_SUITES
- src/lib/ResultsHandler.ts – saveResults + displayResults + performDryRun
- src/lib.ts – thin public API (142 LOC)

All public APIs remain backward compatible. tsc --noEmit passes cleanly.
All pre-existing passing tests continue to pass.

Fixes #52
