+++ src/runners/comprehensive/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Comprehensive UI Test Runner sub-modules
+ */
+
+export { UIFlowTester } from './UIFlowTester';
+export type { TestResults, TabTestFunction } from './UIFlowTester';
