+++ src/runners/smart/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Smart UI Test Runner sub-modules
+ */
+
+export { SmartElementFinder } from './SmartElementFinder';
+export type { UIElement, ElementDiscovery } from './SmartElementFinder';
+export { SmartInteractionExecutor } from './SmartInteractionExecutor';
